move functions into task.rs
