+++ src/main.rs
@@ -22,11 +22,11 @@ fn main() {
             println!("No operation specified. Use --help for more information.");
         }
         Some(op) => match op {
-            Operation::List => fileio::list(),
-            Operation::Show(read_args) => fileio::show(read_args.clone()),
-            Operation::Add(add_args) => fileio::add(add_args.clone()),
+            Operation::List => task::list(),
+            Operation::Show(read_args) => task::show(read_args.clone()),
+            Operation::Add(add_args) => task::add(add_args.clone()),
             Operation::Update(update_args) => todo!("{:?}", update_args),
-            Operation::Remove(read_args) => fileio::remove(read_args.clone()),
+            Operation::Remove(read_args) => task::remove(read_args.clone()),
         },
     }
 }
